Clean and combine code
