--- web/pages/upload-test.tsx
@@ -1,14 +1,0 @@
-// pages/upload-test.js
-
-import { ImageUpload } from "@components/BabyBook/ImageUpload";
-
-const UploadTestPage = () => {
-  return (
-    <div>
-      <h1>Test Image Upload</h1>
-      <ImageUpload />
-    </div>
-  );
-};
-
-export default UploadTestPage;
